Refactoring project commands
