+++ docs/index.md
@@ -10,4 +10,4 @@ Basil CLI is the command-line tool for [Basil](https://github.com/gardenbed/basi
 | `config` | Sets the global configurations for Basil. |
 | `project semver` | Shows the current project [semantic version](https://semver.org). |
 | `project build` | Builds the project in the current directory. |
-| `github release` | Creates a GitHub release with [semantic versioning](https://semver.org). |
+| `project release` | Creates a new release using [semantic versioning](https://semver.org). |
